Ink-Fungus back button complete
Only one bug remains and I have no idea how to fix it: If you hit the back button at the end of a dialog (like, when waiting for input to continue the visual novel), when you load back to the scene it will start at the *next* line of dialog instead of replaying the current line. Again, I don't think this is fixable without wasting time.